Add annotation

+++ homework-2/sql/04.create-user_roles-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE user_roles (
-  user_id INT,
-  role_id INT,
+  user_id BIGINT,
+  role_id BIGINT,
   PRIMARY KEY (user_id, role_id),
   FOREIGN KEY (user_id) REFERENCES users(id),
   FOREIGN KEY (role_id) REFERENCES roles(id)
